% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibInc.R
\name{CalibInc}
\alias{CalibInc}
\title{Prediction of incidence from proxy data using a calibration approach}
\usage{
CalibInc(mod, pred = NULL, weight = 1, aggregate = NULL,
  keep.Vp = FALSE)
}
\arguments{
\item{mod}{Calibration model (type H/I ratio), fitted with
\code{glmer} or \code{glmmPQL}}

\item{pred}{Data frame containing proxy data to be calibrated. If \code{NULL}, try to use the
data.frame used for model evaluation (searched in .GlobalEnv).}

\item{weight}{Weight vector (size 1 for an uniform weight, same size as
\code{nrow(pred)} otherwise). \code{weight} can be given as the name of a
column in \code{pred}.}

\item{aggregate}{Aggregate the predictions according to a combination of variables given as a formula
(see examples). If \code{NULL} aggregation is done according to the combinations of all co-variables
specified in the model formula.}

\item{keep.Vp}{Set to \code{TRUE} to keep variances covariance matrix of the predictions}
}
\value{
A data frame containing predictions values (column \code{pred}) and se (column \code{se})
   at the levels defined in \code{aggregate}. If keep.Vp=\code{TRUE}, the
   variances covariance matrix of the predictions is also returned as an attribute.
}
\description{
\code{CalibInc} functions takes a calibration model and a table with proxy data as
arguments and returns a table with predictions and standard errors
}
\examples{
library(lme4)
library(splines)
library(tidyverse)
data(lopm.CalibSet)
data(lopm.alldist)
## Run the calibration model for lop cancer incidence
## and hospitalisation data in men, 2007-2011
k<-Hmisc::wtd.quantile(lopm.CalibSet$age,weights=lopm.CalibSet$C,p=0.5)\%>\%as.numeric
form.calib<-substitute(H~offset(log(C))+
                       ns(age,knots=k, Boundary.knots = range(age)+c(5,-5))+
                       (1|dist),
                       list(k=k))\%>\%as.formula
mod.calib<-glmer(form.calib,data=lopm.CalibSet\%>\%filter(C>0),family="poisson", nAGQ =20)
## Predict the total number of incident cases by district
CalibInc(mod.calib,pred=lopm.alldist,aggregate=~dist)
## Predict the total number of incident cases by age and district
CalibInc(mod.calib,pred=lopm.alldist,aggregate=~dist+age)
}
\author{
Edouard Chatignoux
}
