% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_fr_95.R
\docType{data}
\name{dep_fr_95}
\alias{dep_fr_95}
\title{French's 95 administrative districts boundaries}
\format{A SpatialPolygons}
\description{
SpatialPolygons of the 95 French's administrative districts.
}
\details{
The Corse départements "2A" and "2B" are regrouped into one single unit "20".
}
\keyword{datasets}
