% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMap.R
\name{ggLegMap}
\alias{ggLegMap}
\title{Internal to add legends to maps}
\usage{
ggLegMap(legend = list(breaks = NULL, density = F, title = NULL), p, color,
  limits)
}
\arguments{
\item{p}{Map in a ggplot format}

\item{color}{Color vector}

\item{limits}{Limits to the maps values}

\item{legDens}{set to  \code{TRUE} to represent distribution of variables as the legend}

\item{legBreak}{Legend breaks}
}
\description{
Internal to add legends to maps
}
\author{
Edouard Chatignoux
}
\keyword{internal}
