% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMap.R
\name{hach.poly}
\alias{hach.poly}
\title{Internal to deal hatching of NS zones}
\usage{
hach.poly(poly, l = 200, shap = -pi/4)
}
\arguments{
\item{poly}{: polygon}

\item{l}{: hatching lines spacement}

\item{shap}{: hatching lines orientation in degree}
}
\description{
Internal to deal hatching of NS zones
}
\author{
Edouard Chatignoux
}
\keyword{internal}
