% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibInc.R
\name{LogNormPI}
\alias{LogNormPI}
\title{Confidence interval for log-normal distributed observations}
\usage{
LogNormPI(data, pred = pred, se = se, level = 0.95)
}
\arguments{
\item{data}{A data frame with log-normal observations}

\item{pred}{Log-normal observation}

\item{se}{Standard error of \code{pred}}

\item{level}{Confidence interval level}
}
\description{
\code{LogNormPI} calcultes confidence interval for log-normal distributed observations
}
\examples{
dt<-data_frame(m=2,sd=1)
LogNormPI(data=dt,m,sd)
}
\author{
Edouard Chatignoux
}
