% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopm.alldist.R
\docType{data}
\name{lopm.alldist}
\alias{lopm.alldist}
\title{LOP cancer in men for French districts.}
\format{A data frame with  1728  rows and  5  variables
\itemize{
\item \code{dist} District
\item \code{age} Central age of the age class
\item \code{H} Number of hospitalizations
\item \code{py} Number of person-years
\item \code{WSP} World Standard Population (Segi 1960 standard)
}}
\usage{
data( lopm.alldist )
}
\description{
Data on hospitalisation for LOP cancer in men over the 2007-2011 period for the 96 French
administrative districts.
}
\details{
The number of hospitalizations and person-years
are tabulated by 5 years age groups (from 0-5 to 90+,
variable `age` corresponds to the central age of the class) and
district (variable `dist`).
}
\references{
http://seer.cancer.gov/stdpopulations/
}
\keyword{datasets}
