% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopm.Fr.R
\docType{data}
\name{lopm.Fr}
\alias{lopm.Fr}
\title{Data on hospitalisation for LOP cancer in men over the 2007-2011 period for the 96 French
adminsitrative districts}
\format{A data frame with  1728  rows and  5  variables}
\usage{
data( lopm.Fr )
}
\description{
\itemize{
\item \code{dist} District
\item \code{age} Central age of the age class
\item \code{H} Number of hospitalizations
\item \code{py} Number of person-years
\item \code{WSP} World Standard Population (Segi 1960 standard)
}
}
\details{
The number of hospitalizations and person-years
are tabulated by 5 years age groups (from 0-5 to 90+,
variable `age` corresponds to the central age of the class) and
district (variable `dist`).
}
\references{
http://seer.cancer.gov/stdpopulations/
}
\keyword{datasets}
