% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMap.R
\name{poly2ggplot}
\alias{poly2ggplot}
\title{Convert a spatialpolygon to a data frame for a ggplot2 use}
\usage{
poly2ggplot(file = NULL, poly = NULL, id = NULL)
}
\arguments{
\item{file}{: File path to shape file}

\item{poly}{: names of a SpatialPolygon if \code{NULL} file}

\item{id}{: id variable for  polygons}
}
\value{
A "polygg" object, to be used with  ggplot2 (ggMap for example)
}
\description{
Convert a spatialpolygon to a data frame for a ggplot2 use
}
\author{
Edouard Chatignoux
}
