% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibInc.R
\name{predict_gam_fixed}
\alias{predict_gam_fixed}
\title{Predict fixed effects part of a gam model with random-effect}
\usage{
predict_gam_fixed(mod, newdata = NULL, type = "lpmatrix")
}
\arguments{
\item{mod}{A model fitted with gam}

\item{newdata}{Prediction data set}

\item{type}{\code{lpmatrix} to get the model matrix, \code{response} for the response}
}
\value{
A matrix or a vector
}
\description{
Predict fixed effects part of a gam model with random-effect
}
\author{
Edouard Chatignoux
}
\keyword{internal}
