% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMap.R
\name{print.polygg}
\alias{print.polygg}
\title{Print a polygg object}
\usage{
\method{print}{polygg}(x)
}
\value{
Print a summary of the polygg object
}
\description{
Print a polygg object
}
\examples{
print(ShapDep)
}
\author{
Edouard Chatignoux
}
\keyword{internal}
